+++ packages/components-react/projects/remix/app/routes/overview.tsx
@@ -1,8 +1,12 @@
 /* Auto Generated File */
-import { PAccordion, PBanner, PButton, PButtonGroup, PButtonPure, PButtonTile, PCarousel, PCheckboxWrapper, PContentWrapper, PDisplay, PDivider, PFieldset, PFlex, PFlexItem, PGrid, PGridItem, PHeading, PHeadline, PIcon, PInlineNotification, PLink, PLinkPure, PLinkSocial, PLinkTile, PLinkTileModelSignature, PMarque, PModelSignature, PorscheDesignSystemProvider, PPagination, PPopover, PRadioButtonWrapper, PScroller, PSegmentedControl, PSegmentedControlItem, PSelectWrapper, PSpinner, PStepperHorizontal, PStepperHorizontalItem, PSwitch, PTable, PTableBody, PTableCell, PTableHead, PTableHeadCell, PTableHeadRow, PTableRow, PTabs, PTabsBar, PTabsItem, PTag, PTagDismissible, PText, PTextareaWrapper, PTextFieldWrapper, PTextList, PTextListItem } from '@porsche-design-system/components-react/ssr';
+import { PAccordion, PBanner, PButton, PButtonGroup, PButtonPure, PCarousel, PCheckboxWrapper, PContentWrapper, PDisplay, PDivider, PFieldset, PFieldsetWrapper, PFlex, PFlexItem, PGrid, PGridItem, PHeading, PHeadline, PIcon, PInlineNotification, PLink, PLinkPure, PLinkSocial, PMarque, PModal, PModelSignature, PorscheDesignSystemProvider, PPagination, PPopover, PRadioButtonWrapper, PScroller, PSegmentedControl, PSegmentedControlItem, PSelectWrapper, PSpinner, PSwitch, PTable, PTableBody, PTableCell, PTableHead, PTableHeadCell, PTableHeadRow, PTableRow, PTabs, PTabsBar, PTabsItem, PTag, PTagDismissible, PText, PTextareaWrapper, PTextFieldWrapper, PTextList, PTextListItem } from '@porsche-design-system/components-react/ssr';
 
 const OverviewPage = (): JSX.Element => {
   const style = `
+    body {
+      overflow: auto !important;
+    }
+
     p-flex-item > p,
     my-prefix-p-flex-item > p,
     p-grid-item > p,
@@ -30,18 +34,16 @@ const OverviewPage = (): JSX.Element => {
       background: deepskyblue;
     }
 
-    p-link-tile,
-    p-button-tile,
-    p-link-tile-model-signature,
-    my-prefix-p-link-tile,
-    my-prefix-p-button-tile,
-    my-prefix-p-link-tile-model-signature {
-      max-width: 400px;
+    .playground--banner,
+    .playground--modal {
+      transform: translate3d(0, 0, 0);
+      height: 10rem;
+      padding: 0;
+      margin: 1rem;
     }
 
-    p-banner,
-    my-prefix-p-banner {
-      --p-banner-position-type: static;
+    .playground--modal {
+      height: 20rem;
     }
   `;
 
@@ -63,11 +65,15 @@ const OverviewPage = (): JSX.Element => {
             <PDisplay>Some Display</PDisplay>
           </div>
 
-          <div className="playground light" title="should render default banner">
+          <div className="playground light playground--banner" title="should render default banner">
             <PBanner>
               <span slot="title">Default banner component</span>
               <span slot="description">Some slotted banner description</span>
             </PBanner>
+          </div>
+
+          <div className="playground light playground--modal" title="should render default modal">
+            <PModal heading="Heading" open={true}>Some Content</PModal>
           </div>
 
           <div className="playground light" title="should render default inline-notification">
@@ -133,6 +139,10 @@ const OverviewPage = (): JSX.Element => {
 
           <div className="playground light" title="should render default fieldset">
             <PFieldset label="Some label" />
+          </div>
+
+          <div className="playground light" title="should render default fieldset-wrapper">
+            <PFieldsetWrapper label="Some label" />
           </div>
 
           <div className="playground light" title="should render default content-wrapper">
@@ -258,20 +268,6 @@ const OverviewPage = (): JSX.Element => {
             <PSwitch>Some label</PSwitch>
           </div>
 
-          <div className="playground light" title="should render default stepper-horizontal">
-            <PStepperHorizontal>
-              <PStepperHorizontalItem state="current">Step 1</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 2</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 3</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 4</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 5</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 6</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 7</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 8</PStepperHorizontalItem>
-              <PStepperHorizontalItem>Step 9</PStepperHorizontalItem>
-            </PStepperHorizontal>
-          </div>
-
           <div className="playground light" title="should render default scroller">
             <PScroller>
               <button>Default</button>
@@ -289,41 +285,6 @@ const OverviewPage = (): JSX.Element => {
               <button>Default</button>
               <button>Default</button>
             </PScroller>
-          </div>
-
-          <div className="playground light" title="should render default link-tile">
-            <PLinkTile href="#" label="Some Label" description="Default">
-              <img
-                src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
-                width={50}
-                height={50}
-                alt="Beach"
-              />
-            </PLinkTile>
-          </div>
-
-          <div className="playground light" title="should render default button-tile">
-            <PButtonTile label="Some Label" description="Default">
-              <img
-                src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
-                width={50}
-                height={50}
-                alt="Beach"
-              />
-            </PButtonTile>
-          </div>
-
-          <div className="playground light" title="should render default link-tile-model-signature">
-            <PLinkTileModelSignature heading="Some Heading" description="Default">
-              <img
-                src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
-                width={50}
-                height={50}
-                alt="Beach"
-              />
-              <PLink slot="primary" href="#">Some label</PLink>
-              <PLink slot="secondary" href="#">Some label</PLink>
-            </PLinkTileModelSignature>
           </div>
 
           <div className="playground light" title="should render default carousel">
@@ -349,11 +310,15 @@ const OverviewPage = (): JSX.Element => {
               <PDisplay>Some Display</PDisplay>
             </div>
 
-            <div className="playground light" title="should render default banner with custom prefix">
+            <div className="playground light playground--banner" title="should render default banner with custom prefix">
               <PBanner>
                 <span slot="title">Prefixed banner component</span>
                 <span slot="description">Some slotted banner description</span>
               </PBanner>
+            </div>
+
+            <div className="playground light playground--modal" title="should render default modal with custom prefix">
+              <PModal heading="Heading" open={true}>Some Content</PModal>
             </div>
 
             <div className="playground light" title="should render default inline-notification with custom prefix">
@@ -422,6 +387,10 @@ const OverviewPage = (): JSX.Element => {
 
             <div className="playground light" title="should render default fieldset with custom prefix">
               <PFieldset label="Some label" />
+            </div>
+
+            <div className="playground light" title="should render default fieldset-wrapper with custom prefix">
+              <PFieldsetWrapper label="Some label" />
             </div>
 
             <div className="playground light" title="should render default content-wrapper with custom prefix">
@@ -547,20 +516,6 @@ const OverviewPage = (): JSX.Element => {
               <PSwitch>Some label</PSwitch>
             </div>
 
-            <div className="playground light" title="should render default stepper-horizontal with custom prefix">
-              <PStepperHorizontal>
-                <PStepperHorizontalItem state="current">Step 1</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 2</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 3</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 4</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 5</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 6</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 7</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 8</PStepperHorizontalItem>
-                <PStepperHorizontalItem>Step 9</PStepperHorizontalItem>
-              </PStepperHorizontal>
-            </div>
-
             <div className="playground light" title="should render default scroller with custom prefix">
               <PScroller>
                 <button>Default</button>
@@ -578,41 +533,6 @@ const OverviewPage = (): JSX.Element => {
                 <button>Default</button>
                 <button>Default</button>
               </PScroller>
-            </div>
-
-            <div className="playground light" title="should render default link-tile with custom prefix">
-              <PLinkTile href="#" label="Some Label" description="Default">
-                <img
-                  src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
-                  width={50}
-                  height={50}
-                  alt="Beach"
-                />
-              </PLinkTile>
-            </div>
-
-            <div className="playground light" title="should render default button-tile with custom prefix">
-              <PButtonTile label="Some Label" description="Default">
-                <img
-                  src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
-                  width={50}
-                  height={50}
-                  alt="Beach"
-                />
-              </PButtonTile>
-            </div>
-
-            <div className="playground light" title="should render default link-tile-model-signature with custom prefix">
-              <PLinkTileModelSignature heading="Some Heading" description="Default">
-                <img
-                  src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADIAAAAyAQMAAAAk8RryAAAABlBMVEUAAAD2vP9xXLiUAAAAAXRSTlMAQObYZgAAABxJREFUGNNjYOBgYGBhYKAZ/R8MDsD4Q5amkz8ASp4PtTYYQZIAAAAASUVORK5CYII="
-                  width={50}
-                  height={50}
-                  alt="Beach"
-                />
-                <PLink slot="primary" href="#">Some label</PLink>
-                <PLink slot="secondary" href="#">Some label</PLink>
-              </PLinkTileModelSignature>
             </div>
 
             <div className="playground light" title="should render default carousel with custom prefix">