+++ packages/utilities/projects/utilities/src/js/grid/gridWidth.ts
@@ -1,7 +1,0 @@
-import { gridWidthMin } from './gridWidthMin';
-import { gridWidthMax } from './gridWidthMax';
-
-export const gridWidth = {
-  min: gridWidthMin,
-  max: gridWidthMax,
-};